Create Utility Conversion Assignment
Assignment is to code a utility that aids in unit conversion. The software should individually prompt user to input four separate lengths, each specified in inches. Following this, the program should convert these lengths and present the results in a table, showing their equivalents in yards, feet, and centimeters. The table should be formatted into four columns, and the values for yards, feet and centimeters should be rounded to two decimal places.